[Feat/#135] 카테고리 생성 후 카테고리 상세로 이동

+++ Staccato-iOS/Staccato-iOS/Service/Category/DTO/PostCategoryResponse.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 struct PostCategoryResponse: Decodable {
-    let categoryId: Int
+    let categoryId: Int64
 }
